add initial clades file
Naming clades in the tree is necessary to allow zoomed-in views on the tree for nextstrain narratives. I had a look through the tree and just looked for clades that I thought would be epidemiologically relevant and worthwhile to be able to describe in more depth in a narrative. Note that clade numberings doesn't mean anything at all, nor are the clades described by any threshold other than that I thought that it would be useful to zoom into the tree at that location.