add BEN234 to dropped_strains file
Sample BEN234 had stochastic alignment issues. MAFFT often struggles when there are large series of N's and BEN234 would sometimes align well, and other times align poorly, with the gap opening shifting the sample's alignment so that it looked like it was highly diverged. Indeed, issue #5 was opened in regards to this sample, which we closed when alignment behavior appeared to be fixed. 

Since this seems to crop up stochastically, we have decided to drop the sample from the build, although all data is still available for anyone seeking to make their own alignments/builds.